+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 **?.?.? Changes:**
 
-* Added 1 new effect:
+* Added 2 new effects:
   * Aspect Roulette: Randomly switches the elite aspect of all characters (only affects players if they already have an aspect equipment)
+  * Unscrap Random Item: Converts a random stack of scrap into a random item of the same tier
 
 * Drop All Items:
   * Fixed effect not working